+++ tests/testthat/test_make_eml.R
@@ -112,6 +112,26 @@ testthat::test_that('Expect argument values in EML', {
       x1$data.table.url[i])
   }
   
+  # data.table.url - A URL is not required for each dataTable
+  
+  x1 <- x
+  x1$data.table.url[1] <- ""
+  r <- do.call(make_eml, x1[names(x1) %in% names(formals(make_eml))])
+  expect_true(
+    is.null(r$dataset$dataTable[[1]]$physical$distribution$online$url))
+  expect_equal(
+    r$dataset$dataTable[[2]]$physical$distribution$online$url,
+    x1$data.table.url[2])
+  
+  x1 <- x
+  x1$data.table.url[1] <- NA_character_
+  r <- do.call(make_eml, x1[names(x1) %in% names(formals(make_eml))])
+  expect_true(
+    is.null(r$dataset$dataTable[[1]]$physical$distribution$online$url))
+  expect_equal(
+    r$dataset$dataTable[[2]]$physical$distribution$online$url,
+    x1$data.table.url[2])
+  
   # other.entity
   
   x1 <- x
@@ -150,6 +170,26 @@ testthat::test_that('Expect argument values in EML', {
       r$dataset$otherEntity[[i]]$physical$distribution$online$url,
       x1$other.entity.url[i])
   }
+  
+  # other.entity.url - A URL is not required for each otherEntity
+  
+  x1 <- x
+  x1$other.entity.url[1] <- ""
+  r <- do.call(make_eml, x1[names(x1) %in% names(formals(make_eml))])
+  expect_true(
+    is.null(r$dataset$otherEntity[[1]]$physical$distribution$online$url))
+  expect_equal(
+    r$dataset$otherEntity[[2]]$physical$distribution$online$url,
+    x1$other.entity.url[2])
+  
+  x1 <- x
+  x1$other.entity.url[1] <- NA_character_
+  r <- do.call(make_eml, x1[names(x1) %in% names(formals(make_eml))])
+  expect_true(
+    is.null(r$dataset$otherEntity[[1]]$physical$distribution$online$url))
+  expect_equal(
+    r$dataset$otherEntity[[2]]$physical$distribution$online$url,
+    x1$other.entity.url[2])
   
   # provenance - Get provenance metadata for EDI data packages and place under
   # /eml/dataset/methods/methodStep/dataSource